requirements.txt
aiohttp
discord.py
python-decouple